api/requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Werkzeug==2.0.3
uWSGI==2.0.20
python-dotenv==0.19.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
flask_bcrypt==0.7.1
psycopg2==2.9.3